long

+++ geolocator_android/android/src/main/java/com/baseflow/geolocator/location/BackgroundNotification.java
@@ -87,7 +87,7 @@ public class BackgroundNotification {
                 .setContentText(options.getNotificationText())
                 .setContentIntent(buildBringToFrontIntent())
                 .setOngoing(options.isSetOngoing())
-                .setColor(options.getNotificationColor()).setColorized(true);
+                .setColor(options.getNotificationColor().intValue()).setColorized(true);
 
         if (notify) {
             NotificationManagerCompat notificationManager = NotificationManagerCompat.from(context);
